use fixed delta time
The original  system we are porting from used the fixed delta time in the Bee Manager.  Any system that is porting from beemanager needs to also used fixed delta to get something that looks like the original

I really dislike that we are mixing up delta & fixed delta time like this.